feat: Add Movie page and ModalReview component
The code changes include adding a new page called "Movie" and a new component called "ModalReview". The "Movie" page is imported in the App component and rendered as a route. The "ModalReview" component is a modal that allows users to add reviews.

This commit message follows the established convention of using a prefix to indicate the type of change ("feat" for a new feature) and provides a clear and concise description of the changes made.